Add rofi_highlight
This adds the ability to use rofi's highlighting capabilities, that
were introduced in rofi 0.15.4. It has to be enabled by the user.
So existing users, do not see a change in behaviour. If it is enabled,
the active connections are no longer marked with '**' in the beginning,
but highlighted by rofi (in a color depending on the user's
configuration).