Add a hint to check if NetworkManager is running
I'm not sure this is the best way to detect if the NetworkManager
is running but it seems to be the more global way to detect it.

Also I don't raise an error but prefer to print a warning message
to avoid the program to crash while the NetworkManager is running.